requirements.txt
# pyLIQTR package requirements.

# Google originated
# October 2023 release currently required by qualtran 0.2.0
# cirq-core==1.3.0.dev20231018023458
# removing the restriction and letting Qualtran add their restriction(s) and the wheel will figure it out
cirq-core
# November 2023 release we were previously using
# cirq-core==1.3.0.dev20231102230836
qualtran==0.4.0

# Other requirements
ipykernel
numpy
pandas
scipy<1.13.0
matplotlib
tqdm
openfermion
portalocker
ply
gmpy2
networkx
juliacall
juliapkg
rustworkx
# Removed the following from the install because it is not yet available on Windows
# openfermionpyscf
